Remove VisualStudioDogfoodSetup
This now was only being used to register a trace listener. Moved it to the managed package and #if it out so that it's only enabled under DEBUG.
