+++ src/ProjectSystem.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26226.0
+VisualStudioVersion = 15.0.26228.4
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeployTestDependencies", "DeployTestDependencies\DeployTestDependencies.csproj", "{37BA82E6-9ABD-4ACA-AA26-2DFD39A359A5}"
 EndProject
@@ -56,8 +56,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Targets", "Targets", "{573B
 		Targets\Microsoft.Managed.DesignTime.targets = Targets\Microsoft.Managed.DesignTime.targets
 		Targets\Microsoft.VisualBasic.DesignTime.targets = Targets\Microsoft.VisualBasic.DesignTime.targets
 	EndProjectSection
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ProjectSystemDogfoodSetup", "ProjectSystemDogfoodSetup\ProjectSystemDogfoodSetup.csproj", "{F9522F27-B9AC-4D6F-8F54-6A8BD4D33574}"
 EndProject
 Project("{778DAE3C-4631-46EA-AA77-85C1314464D9}") = "SettingsDesignerUnitTests", "Microsoft.VisualStudio.Editors.SettingsDesigner.UnitTests\SettingsDesignerUnitTests.vbproj", "{A03E115A-9193-455D-A70A-CFEB943D0319}"
 EndProject
@@ -146,10 +144,6 @@ Global
 		{369DCA0F-A079-4D0C-8B32-53879199B681}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{369DCA0F-A079-4D0C-8B32-53879199B681}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{369DCA0F-A079-4D0C-8B32-53879199B681}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F9522F27-B9AC-4D6F-8F54-6A8BD4D33574}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F9522F27-B9AC-4D6F-8F54-6A8BD4D33574}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F9522F27-B9AC-4D6F-8F54-6A8BD4D33574}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F9522F27-B9AC-4D6F-8F54-6A8BD4D33574}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A03E115A-9193-455D-A70A-CFEB943D0319}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A03E115A-9193-455D-A70A-CFEB943D0319}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A03E115A-9193-455D-A70A-CFEB943D0319}.Release|Any CPU.ActiveCfg = Release|Any CPU
