--- src/ProjectSystemDogfoodSetup/Properties/launchSettings.json
@@ -1,9 +1,0 @@
-{
-  "profiles": {
-    "ProjectSystemDogfoodSetup": {
-      "commandName": "Executable",
-      "executablePath": "$(DevEnvDir)devenv.exe",
-      "commandLineArgs": "/rootsuffix RoslynDev /log"
-    }
-  }
-}
